[src] Renamed .paths to .exports [src] Updated README with recent change [src] Added GOROOT v1.16.6 and set as default [src] Added folder name validation before removing the .dotfiles directory

+++ dotfiles/dotfiles.sh
@@ -105,7 +105,7 @@ uninstall_dotfiles() {
   _unlink_managed_files_in_dotfiles_dir
   _unlink_custom_files_in_dotfiles_dir
 
-  if [[ ${DOTFILES_HOME_DIR} != ${HOME} ]]; then
+  if [[ ${DOTFILES_HOME_DIR} != ${HOME} && -d ${DOTFILES_HOME_DIR} && ${DOTFILES_HOME_DIR} == *"dotfiles"* ]]; then
     rm -r ${DOTFILES_HOME_DIR}
     output+="\nCleanup:\n  ${DOTFILES_HOME_DIR} --> removed."
   fi
